backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0
gunicorn==21.2.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
requests==2.31.0
fuzzywuzzy==0.18.0
phonetics>=1.0.0
nltk>=3.8.1
textblob>=0.17.1
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4
spacy>=3.7.0
networkx>=3.1
plotly>=5.15.0
dash>=2.14.0
redis>=4.5.0
celery>=5.3.0
pytest>=7.4.0
statsmodels>=0.14.0
prophet>=1.1.5
